% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name.R
\name{describe}
\alias{describe}
\title{Get a summary description of a first name}
\usage{
describe(name)
}
\arguments{
\item{name}{first name.}
}
\value{
A character vector summarising statistics about the first name.
}
\description{
Get a summary description of a first name
}
\examples{
describe("Astrid")
}
