% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name.R
\name{get_count}
\alias{get_count}
\title{Determine the frequency count of a first name in Estonia}
\usage{
get_count(name)
}
\arguments{
\item{name}{first name.}
}
\value{
A numeric vector showing the number of people having such a name.
}
\description{
Determine the frequency count of a first name in Estonia
}
\examples{
get_count("Astrid")
}
