% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name.R
\name{get_count_by_month}
\alias{get_count_by_month}
\title{Determine the frequency count of a first name in Estonia by birth month}
\usage{
get_count_by_month(name)
}
\arguments{
\item{name}{first name.}
}
\value{
A numeric vector showing the number of people having such a name by birth month.
}
\description{
Determine the frequency count of a first name in Estonia by birth month
}
\examples{
get_count_by_month("Astrid")
}
