% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name.R
\name{get_gender}
\alias{get_gender}
\title{Determine the gender of a first name}
\usage{
get_gender(name)
}
\arguments{
\item{name}{first name.}
}
\value{
The gender of the name - either \code{male}, \code{female} or \code{both}.
}
\description{
Determine the gender of a first name
}
\examples{
get_gender("Astrid")
}
