% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name.R
\name{get_normalized_count_by_county}
\alias{get_normalized_count_by_county}
\title{Determine the frequency count of a first name in Estonia by county}
\usage{
get_normalized_count_by_county(name)
}
\arguments{
\item{name}{first name.}
}
\value{
A numeric vector showing the number of people having such a name out of 10,000 people by county.
}
\description{
Determine the frequency count of a first name in Estonia by county
}
\examples{
get_count_by_county("Astrid")
}
