% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name.R
\name{get_rank}
\alias{get_rank}
\title{Determine the rank of a first name in Estonia}
\usage{
get_rank(name)
}
\arguments{
\item{name}{first name.}
}
\value{
A numeric vector showing the popularity rank among female and male first names.
}
\description{
Determine the rank of a first name in Estonia
}
\examples{
get_rank("Astrid")
}
